commit 9554f7e18d034b915acbd51b3c4e3261c88e166d
    12094 loader should be configured for serial console
commit 6644025c7c20579199e62886be7da2712f2f213b
    11641 spelling mistakes in section 7d of the manual
commit f3065a883c2d0dfcd5ac6f21692e95c945b23d2c
    11639 some man pages show incorrect driver locations
commit 1d77dcda79ea31b52efaac166328f8a7f7359d5a
    12138 Want smbios pointer device decoding
commit 176a9270a1856e0e2fddcbc3c7d35aa1f750c6b9
    12133 Want support for SMBIOS 3.3
commit b63b86d12b5ac4c5450c67c1222a35f224f5d0e4
    12158 libdtrace: argument to 'alloca' may be too large
commit 2c687d6826d5cbe1435931df7165d210c3336019
    12131 mdb: cast between incompatible function types
commit 2cac0f114247a29fba32510c4c0ea549c8c0a7c9
    12106 libshell: cast between incompatible function types
commit 8bbe90c14cec0e58886bcb1d4d11723245046f36
    12109 libfksmbfs: cast between incompatible function types
commit 164ffd8939460c53fad1cb0c2a860ad54badd3e1
    12164 audio: definition of implicit copy constructor/operator is deprecated

+++ usr/src/cmd/mdb/common/kmdb/kmdb_kvm.c
@@ -2394,9 +2394,9 @@ kmt_destroy(mdb_tgt_t *t)
 
 static const mdb_tgt_ops_t kmt_ops = {
 	kmt_setflags,				/* t_setflags */
-	(int (*)()) mdb_tgt_notsup,		/* t_setcontext */
+	(int (*)())(uintptr_t) mdb_tgt_notsup,	/* t_setcontext */
 	kmt_activate,				/* t_activate */
-	(void (*)()) mdb_tgt_nop,		/* t_deactivate */
+	(void (*)())(uintptr_t) mdb_tgt_nop,	/* t_deactivate */
 	kmt_periodic,				/* t_periodic */
 	kmt_destroy,				/* t_destroy */
 	kmt_name,				/* t_name */
@@ -2425,25 +2425,25 @@ static const mdb_tgt_ops_t kmt_ops = {
 	kmt_addr_to_ctf,			/* t_addr_to_ctf */
 	kmt_name_to_ctf,			/* t_name_to_ctf */
 	kmt_status,				/* t_status */
-	(int (*)()) mdb_tgt_notsup,		/* t_run */
+	(int (*)())(uintptr_t) mdb_tgt_notsup,	/* t_run */
 	kmt_step,				/* t_step */
 	kmt_step_out,				/* t_step_out */
 	kmt_next,				/* t_next */
 	kmt_continue,				/* t_cont */
-	(int (*)()) mdb_tgt_notsup,		/* t_signal */
+	(int (*)())(uintptr_t) mdb_tgt_notsup,	/* t_signal */
 	kmt_add_vbrkpt,				/* t_add_vbrkpt */
 	kmt_add_sbrkpt,				/* t_add_sbrkpt */
 	kmt_add_pwapt,				/* t_add_pwapt */
 	kmt_add_vwapt,				/* t_add_vwapt */
 	kmt_add_iowapt,				/* t_add_iowapt */
-	(int (*)()) mdb_tgt_null,		/* t_add_sysenter */
-	(int (*)()) mdb_tgt_null,		/* t_add_sysexit */
-	(int (*)()) mdb_tgt_null,		/* t_add_signal */
+	(int (*)())(uintptr_t) mdb_tgt_null,	/* t_add_sysenter */
+	(int (*)())(uintptr_t) mdb_tgt_null,	/* t_add_sysexit */
+	(int (*)())(uintptr_t) mdb_tgt_null,	/* t_add_signal */
 	kmt_add_trap,				/* t_add_fault */
 	kmt_getareg,				/* t_getareg */
 	kmt_putareg,				/* t_putareg */
-	(int (*)()) mdb_tgt_nop,		/* XXX t_stack_iter */
-	(int (*)()) mdb_tgt_notsup		/* t_auxv */
+	(int (*)())(uintptr_t) mdb_tgt_nop,	/* XXX t_stack_iter */
+	(int (*)())(uintptr_t) mdb_tgt_notsup	/* t_auxv */
 };
 
 /*
